Fixes the RBMK having an infinite meltdown state + balance. (#1038)
## About The Pull Request


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/46219e62-5232-4f28-a1e4-d1c099ef1b3d)


## Proof Of Testing

I did not test this but it will work. This is my burger seal of
approval.


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: BurgerBB
balance: The RBMK can now handle more power generation before going into
meltdown with upgraded parts, and also generates less power with
upgraded parts.
fix: Fixes the RBMK having an infinite meltdown state when safeties
trigger at the same time as a meltdown.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->