Exempt playtime by job basis (#1135)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Adds functionality for administrators to toggle playtime job exemption
for ckeys on a job basis
Admins go through player panel to access the menu and they can toggle
that with just one click
The action is logged, announced to all admins, and if the player is
logged on with that ckey, it is announced to him aswell
The exemption status is saved on the DB, a new `bubber_schema.sql` needs
to be ran to add the new table to the DB

![dreamseeker_gHffQnzd8E](https://github.com/Bubberstation/Bubberstation/assets/6381979/b726c730-45fd-4dac-a84b-df0a81dd375f)

![dreamseeker_wTJk1mZnTg](https://github.com/Bubberstation/Bubberstation/assets/6381979/b639fab4-69f1-4407-ba54-e015bb00a713)

![dreamseeker_MvNrETfYGk](https://github.com/Bubberstation/Bubberstation/assets/6381979/dbe6c384-2a00-4e6e-bb0b-56c311fc261a)

![heidisql_6YRaRhl2g3](https://github.com/Bubberstation/Bubberstation/assets/6381979/bd307ca6-be10-4da4-8e07-d6d431f4fa95)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
It was requested

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

Included screenshots in the About section

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
admin: administrators can now toggle job playtime exemption by job basis
for players
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>