Replaces the Meteor Shuttle with THE CUBE (#1171)
## About The Pull Request

Replaces the Meteor Shuttle with THE CUBE. It's an emag only emergency
shuttle that is literally just a 16x16 cube with some engines on it.


## Why It's Good For The Game

Asteroid Shuttle had a lot of issues with it actually being
spawned/launche correctly so I'm removing it and replacing it with this.
It is less destructive, but still very annoying and worthy of being an
emag shuttle.

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/99b1c49d-68bd-46cd-91a8-cac0ea980ce0)



## Changelog


:cl: BurgerBB
del: Replaces the Meteor Shuttle with THE CUBE. It's an emag only
emergency shuttle that is literally just a 16x16 cube with some engines
on it.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>