RULE 1: Always keep the muzzle pointed in a safe direction (#1223)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes gun safeties. BETTER FOLLOW RULE 1 OF GUN SAFETY BETTER NOW!

This is actually extremely noninvasive as far as uhhh uhhh uhhh
modularity goes because it's just ONE file, all I had to do was disable
the add component part of the proc but the proc is still defined so it
doesn't error out on the guns that get the proc in other files.

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

I just find them really annoying.

Also the bigger thing, there's a very strong supermajority for removing
it in the
[suggestions](https://discord.com/channels/1059199070016655462/1071095123145924679/1212230353561849857)
channel.


![image](https://github.com/Bubberstation/Bubberstation/assets/26744576/4c8228ca-72de-4690-a09f-a38d2fda881c)


## Proof Of Testing

I loaded in, joined as captain, picked up all the guns in the armory and
didn't see safeties on ANY of them.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:Swiftfeather
del: Removed gun safeties.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>