Lunchboxes (#1020)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds lunchboxes, available in your loadout under "other". It can
only hold food and drinks, and it's limited to 4 slots so you can't
stuff it with 12 big bite burgers and never visit the kitchen again.

There's seve- ...uh, _**six**_ options to pick from, those being -
- Nanotrasen
- Medical
- Safety Moth
- Hearts
- Corgi
- Rabbit
- the other one

## Why It's Good For The Game

It adds cute RP stuff, you can pack lunches for your friends, you can
take your food to go if you're in a hurry - but not so much food that
you don't need to go to the kitchen anymore.

## Proof Of Testing


![lunchbox](https://github.com/Bubberstation/Bubberstation/assets/58074918/289866b7-dab4-4618-b218-dcfe93016a5b)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: adds lunchboxes that hold food and drinks to the loadout
image: added lunchbox sprites
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>