Fixes some Moon Station bugs, such as the funny engineering bridge access bug and a lack of cigarette/coffee vendors. (#1188)
## About The Pull Request

Fixes some Moon Station bugs, such as the funny engineering bridge
access bug and a lack of cigarette/coffee vendors.

## Why It's Good For The Game

Bugfixes good.

## Proof Of Testing

If it compiles, it werks.

## Changelog

:cl: BurgerBB
fix: Fixes some Moon Station bugs, such as the funny engineering bridge
access bug and a lack of cigarette/coffee vendors.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>