Nerfs Lichdom because someone messaged me on discord about it once (#1078)
## About The Pull Request

After 3 Lichdom resurrections, announcements will be made about your
general phylactery location (the area) when you die.
Phylactery binding now requires the object to be at least bulky in size,
and it cannot be fireproof, lava proof, or indestructible.

## Why It's Good For The Game

Lichdom is insanely strong, even with the drawbacks that make you
generally weaker. Items can be hidden away in space (skeletons are
immune to space!) and can be small or hard to see objects such as pens
or cigarette butts. Some items can be fireproof and acid proof as well,
requiring usage of mechanics that delete objects entirely (very limited
and easy to sabotage, maybe the supermatter is reliable but that can be
sabotaged too).

Regardless this PR should make lichdom less crazy.

## Proof Of Testing

Untested and in a draft because feedback is needed.

## Changelog

:cl: BurgerBB
balance: After 3 Lichdom resurrections, announcements will be made about
your general phylactery location (the area) when you die. Phylactery
binding now requires the object to be at least bulky in size, and it
cannot be fireproof, lava proof, or indestructible.
/:cl:

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>