Stop spamming logs when engineers lovingly feed the supermatter pipes by the hundreds. (#1116)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Title 

Also the method I went about this adds a flag to the game, which means
in the future if people find another way to spew even more items into
the supermatter that spam the log, we can just add the flag to those
items.


## Why It's Good For The Game

I got bwoinked this one time because I was spamming the admin log and I
told them I'd look for a solution.

This might not be _the_ solution but it is **a** solution. 

## Proof Of Testing



this is what my IC chat tab looked like

![image](https://github.com/Bubberstation/Bubberstation/assets/26744576/bf7e5b9e-daf5-451c-b839-3232ae0a6317)





This is what my logs tab looked like


![image](https://github.com/Bubberstation/Bubberstation/assets/26744576/5c820074-a16e-47d1-83cc-606fe6874eb0)


It still logs pipes for initial startup so we good there.
I also verified that other items still set off the log function as
expected.

## Changelog


:cl:
admin: Will no longer receive supermatter ate a pipe spam in the log.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>