Blood filtration surgery removes toxins (#1206)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
The basic premise of this is that bloodfiltering removes a small amount
of toxin damage, to prevent this being abusable it will only repeat when
they have reagents in their system. It works on slimes, however the
trade off is that it's worse than medications (80% seiver at 1000k is
like... 6 toxin per second, this is about 2.5 per second if my math is
right) and will purge any meds in the subjects sytem. (Even the
beneficial ones, use it wisely)

Currently; You _can_ heal toxins via surgery, however the stomach pump
surgery cannot be used on many races, some of which are very commonly
used by our players.
<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
If you're filtering someone's blood, surely cycling clean blood around
their system would purge some of the toxins (since that's kinda what our
bloodstream does, in combination with our liver and kidneys). It is also
convenient for those times where you are out of medications (Blackouts,
breaches or it's otherwise unsafe to get them) and need to heal
someone's toxins... click by click if required.

It's also much more humane than wrenching a carbon's stomach to make
them sick and can work on slimes/other species that aren't listed as
carbon.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing

![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/cbf493af-d892-4b1c-973e-4483b9aa75c5)

![image-1](https://github.com/Bubberstation/Bubberstation/assets/53197594/73321fdf-85a1-4720-99d8-4b05c0b94160)


<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Blood filtering now restores toxins.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>