[DONE] Adds the weakness to sol quirk (#1255)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
You know how bloodsuckers are weak to Sol? Now you can be too! Only as a
hemophage.
Features included: The very same warnings and UI that bloodsuckers get
to know when they are in danger!
This quirk also makes healing in a coffin cost only half blood, neat
huh?
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
People wanted to suffer, and why am I to deny them pain?
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing
HUD/ burning / healing works fine
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Adds the hemophage-only quirk, Sol weakness
code: Refactors the sol subsystem so that itself decides if it should
run or not based on if anyone is using it.
code: Refactors bloodsucker UI elements so that they can function fine
outside a bloodsucker antag datum
code: bloodsucker_blood_volume, humanity_lost, bloodsucker_level and
bloodsucker_level_unspent are now private variables due to them being
used for UI updates, if you need to change them, use the procs that are
written in the autodoc
code: Quirk species whitelists no longer use bitflags, but use the
species ID's. Easily allowing you to add new whitelists without having
to mess around with making a new bitfield
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>