Removes Dauntless Self Destruct (#1285)
## About The Pull Request

Removes the Dauntless self destruct and the dense rock surrounding the
base. Sets the exterior turrets to neutral.

## Why It's Good For The Game

No one uses it anyway and I've come to the conclusion that it's not
really needed. The dense rock primarily existed to stop a stray colossus
bolt from detonating the entire base. With the removal of the dense
rock, the exterior turrets will only target fauna except these two
hanger turrets (these will still shoot you)


![image](https://github.com/Bubberstation/Bubberstation/assets/95130227/d33248be-1249-4c55-9e2e-fccb18323c2e)


Also it's not like it's not super easy to cheese the already existing
turrets with the dense rock or a few selective mined out rocks. They're
toned back signifigantly because Dauntless is chunky, and I'm not going
to make ~ 60 x 70 area of lavaland (base + turret range) automatic kill
on sight if you accidently stray a bit too close to the three shot
turrets.

## Proof Of Testing

It compiles and works on the live server.

## Changelog

:cl:
del: Removes Dauntless self destruct and dense rock surrounding the base
balance: Dauntless exterior turrets shoot fauna, not people
/:cl:

---------

Co-authored-by: Return <donwest947@gmail.com>