[SALT PR] Death bolts no longer dust corpses (#1320)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
What it says on the tin, again. 
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game
Dying to a death bolt, or getting hit with one while dead (Which isn't
hard given they're specifically coded to hit corpses) normally dusts
you, which is the one remaining instance of megafauna being a hard round
removal. Some fauna, such as vileworm or legion, eat you, but those
still allow recovery if the mob is killed, whereas even if a daring
miner went to kill the wendigo that downed you, and you didn't get
dusted already, you're almost guaranteed to be gone from a death bolt by
the time the fight's over.

Add to that the complete lack of feedback that it was in fact a death
bolt that did it, and it can lead to people being dusted with no
explanation when fighting colossus/wendigo.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing
Spawned 3 wendigos and a colossus, got nuked by death bolt, corpse still
here. 👍
<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:VrazzleDazzle
del: Death bolts no longer dust you on death.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->