Doctor Moffson Antivirus for Artificial Intelligence Units @ the AI restore consoles (#186)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR makes it so that you can fix your malfunctional (naughty) AI
instead of wiping and RR'ing them.
Yes, to get to this point you need to defeat the AI. Keep the player in
the round, everyone benefits, etc.

https://cdn.discordapp.com/attachments/610319137012580383/1049887877972099083/dreamseeker_QRntph4a7Z.mp4
## How This Contributes To The Skyrat Roleplay Experience

Lets players do more without the fear of sitting in ghost and getting
jaded so fast by the inevitable "malf bad" gang busting your doors down

![SftFpIRyQm](https://user-images.githubusercontent.com/77420409/206078841-b7df3fce-9b8c-4906-8abd-823faf41634e.png)

## Proof of Testing

<!-- Include any screenshots/videos/debugging steps of the code
functioning successfully, between the </summary> and </details> code
blocks. -->

<details>
<summary>Screenshots/Videos</summary>
  


https://user-images.githubusercontent.com/77420409/206078823-826e331b-8242-49e5-abfa-ad774ea8fd75.mp4



https://user-images.githubusercontent.com/77420409/206079197-d8e38bed-609c-467f-b6e0-4776f55be4c2.mp4


</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

🆑 
add: You can now un-malf your malf-AI!
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: iero-srv <>