Moonstation Misc Bugfixes (#1317)
## About The Pull Request

Fixes the following moonstation bugs

- Fixes safes not storing their contents in the vault.
- Fixes random-ass safe in the security lobby.
- Redoes disposals a little bit so the conveyor has to be on in order to
shred objects (instead of being thrown in).
- Fixes the gateway being the wrong subtype.


## Why It's Good For The Game

a

## Proof Of Testing

a

## Changelog

:cl: BurgerBB
fix: Fixes a few moonstation bugs.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>