[TG MIRROR] Fixes DRAGnet net mode not dealing stamina damage (#1343)
## [OriginalPR](https://github.com/tgstation/tgstation/pull/82320)

fixes #1067 

## About The Pull Request

This was meant to be a thing. Looking at the gitblame it has been broken
for **9 fucking years** (Ever since it was added):



![image](https://github.com/tgstation/tgstation/assets/49160555/5c3d98fa-32f5-4fab-9b9d-301905c5c846)

How the hell did NOBODY notice for so fucking long

## Why It's Good For The Game

Actually makes something work as intended. Might make sec consider less
lethal options when dealing with threats.