Removes Lips bodymarking east and west sprite  (#1347)
## About The Pull Request

Removes the east and west sprite for the Lips body marking

## Why It's Good For The Game

To be quite honest, it looks quite ugly every time you turn your sprite
left or right since it's a pixel that overlaps with the sprite outline.
It doesn't need an east or west. Lips are a very minor, single pixel
thing.


![image](https://github.com/Bubberstation/Bubberstation/assets/95130227/fdb10086-a7d2-4436-a6c9-2036ac60185e)


## Proof Of Testing

I deleted two pixels. 

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
image: Removes the east and west sprite on the Lips bodymarking
/:cl:


<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->