Lavaland Dauntless tweaks (#1301)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR tweaks the layout for the lavaland Dauntless base. It does the
following:
* Expand the hanger by 1 unit so blackmarket trader shuttle can dock
without clipping onto distro pipes
* Redo dorms from scratch so two dorm rooms have lava windows (complete
with privacy shutters)
* Tweak layout for kitchen and hydro so they are not as cramped
* Tweak layout for medical to have more gawking area (it is a common
place to congregate)
* Readded Insurgent Care healthbot from DS-2 into medbay
* Wrench down R&D console in research
* Add syndie soap bar to showers

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

This is directly in response to feedback from players who regularly play
on Dauntless.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof Of Testing


![preview](https://github.com/Bubberstation/Bubberstation/assets/80724828/97aee5f4-fe3b-4535-9b0c-417df757001a)

This is an older screenshot composite of the new layout.
**Please ignore the dense rock, jukebox, extra crematorium and holofan
for the turbine** - those were tests and are not included in this PR.


<img width="579"
src="https://github.com/Bubberstation/Bubberstation/assets/80724828/065db37d-c418-49ed-af21-5ed6c8132345">

This is the blackmarket shuttle fitting inside neatly.

<!-- Compile and run your code locally. Make sure it works. This is the
place to show off your changes! We are not responsible for testing your
features. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
qol: Tweak layout of Dauntless kitchen, dorms and medbay
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>