Luxury Sauna Pod (#1350)
## About The Pull Request

Nanotrasen has given miners the luxury sauna pod for all your sauna
deployment needs. I had this idea on the backburner for a long while,
and I finally decided to just impliment it.

## Why It's Good For The Game

I have a major sauna bias. But I also think we could have things in
mining that are a little pricy, and requires some materials and ore to
engage with.

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/95130227/d8107704-a747-40b6-a115-f6e004c9195e)


## Changelog

:cl:
add: Sauna survival pods. Found in mining vendor and occasionally
maintenance
/:cl: