Allows Ashwalkers to Breathe on Station (#1297)
## About The Pull Request

Allows Ashwalkers to Breathe on Station.

Previously, possibly due to an oversight, ashwalkers required nitrogen
(and sometimes plasma) in the air in order to breathe it.

## Why It's Good For The Game

QoL, basically. Prevents Ashwalkers from gasping constantly even though
they can survive the air on station.

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/8625ced7-9d08-43f0-9c89-e1602844431f)


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/efe09ee6-9bef-4075-b164-26a88df1f68f)


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/39665221-351f-47a7-b83b-3cd54854a255)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: BurgerBB
qol: Allows Ashwalkers to Breathe on Station
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>