Ports BoxStation (NSSJourney) from FluffySTG (#1128)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
A nostalgic return of Boxstation with updates.

https://www.youtube.com/watch?v=guBsPkUkPQg

![box3](https://github.com/Bubberstation/Bubberstation/assets/96078776/ae8cfd5f-1879-42bc-a459-455b4b44b3db)

Derived from FluffySTG, includes minimal Bubber changes.

## Why It's Good For The Game

Gives an alternative option for a Box-type station without the presence
of ice.

## Proof Of Testing

Compiles. Runs. Full round tested.
Currently gathering feedback and maintaining.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl: Shayoki
add: Adds Boxstation
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>
Co-authored-by: Waterpig <49160555+Majkl-J@users.noreply.github.com>
Co-authored-by: Cyprex <35031555+Cyprex@users.noreply.github.com>