Adds lifespan to Endurogrow (#1379)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This PR adds lifespan to endurogrow, meaning you can get pure chemical
reagents without grafting, since they require potency, endurance,
lifespan and nil instability for absolute purity. While quicker,
grafting requires more trays and seeds so that's the primary trade off,
alongside the fact that endurogrow can't be used alongside a
diethy-saltpetre mix because it butches potency and yield while in the
trays.

EDIT: Just because I'm a bit of a sadist and didn't want it to be too
easy. I've completely removed the ability to get harvest while
bolstering endurance/lifespan. I've tripled the negatives of endurogrow
so you have to commit to use it for it's benefits.

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

Purer chems without needing 6 trays just growing olives or harebells for
grafting.

## Proof Of Testing


![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/ba7b254b-01b3-42e6-947f-b7f7792fbf42)

![image](https://github.com/Bubberstation/Bubberstation/assets/53197594/9c55ee11-ba00-422c-b7fa-57f97f8068b0)


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
qol: made endurogrow boost lifespan.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

<!-- By opening a pull request. You have read and understood the
repository rules located on the main README.md on this project. -->

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>